+++ lib/features/login/model/kakao_login_request_model.dart
@@ -1,0 +1,13 @@
+class KakaoLoginRequestModel{
+  final String accessToken;
+
+  KakaoLoginRequestModel({
+    required this.accessToken
+  });
+
+  Map<String, dynamic> toJson() {
+    return {
+      'accessToken': accessToken
+    };
+  }
+}
