+++ lib/features/login/model/login_request_model.dart
@@ -16,16 +16,3 @@ class LoginRequestModel {
   }
 }
 
-class KakaoLoginRequestModel{
-  final String accessToken;
-
-  KakaoLoginRequestModel({
-    required this.accessToken
-  });
-
-  Map<String, dynamic> toJson() {
-    return {
-      'accessToken': accessToken
-    };
-  }
-}
